update function demo cases
