+++ pytest.ini
@@ -6,7 +6,3 @@ markers =
 
     satelliteOnly: cases for satellite
     rhsmOnly: cases for candlepin
-
-    packageInfo: base cases to check virt-who package info
-    multiHypervisor: cases to report multi hypervisors together
-
