+++ tests/function/test_config.py
@@ -42,3 +42,20 @@ class TestGlobalConfiguration:
             1.
         """
         logger.info("Succeeded to run the 'test_interval_in_virtwho_conf'")
+
+    @pytest.mark.tier3
+    def test_oneshot_in_virtwho_conf(self):
+        """Just a demo
+
+        :title: virt-who: config: test oneshot option
+        :id: 15b45691-7358-4bc8-b49f-f87100bded1b
+        :caseimportance: High
+        :tags: tier3
+        :customerscenario: false
+        :upstream: no
+        :steps:
+            1.
+        :expectedresults:
+            1.
+        """
+        logger.info("Succeeded to run the 'test_oneshot'")
